+++ mango3-studio/src/components/mod.rs
@@ -7,13 +7,11 @@ use mango3_leptos_utils::models::WebsiteResp;
 use crate::context::use_my_website_resource;
 
 mod my_websites_infinite_scroll;
-mod page_form_fields;
 mod post_form_fields;
 mod selected_website_dropdown;
 mod theme_selector_field;
 
 pub use my_websites_infinite_scroll::MyWebsitesInfiniteScroll;
-pub use page_form_fields::PageFormFields;
 pub use post_form_fields::PostFormFields;
 pub use selected_website_dropdown::SelectedWebsiteDropdown;
 pub use theme_selector_field::ThemeSelectorField;